refactor: 면접 API 중복 컨텍스트 로딩 헬퍼로 추출
